some refactor v3

+++ internal/models/user_balance.go
@@ -2,5 +2,5 @@ package models
 
 type UserBalance struct {
 	Current   float32 `json:"current" db:"balance"`
-	Withdrawn int     `json:"withdrawn" db:"withdrawn"`
+	Withdrawn float32 `json:"withdrawn" db:"withdrawn"`
 }
